handle submodule updates and url changes
If a submodule had an update, it would show up as a + in the 'git submodule
status' output, and thus would not be automatically updated because we were only
looking for '-', so the regexp was changed to look for + as well as -

Additionally, add a 'git submodule sync' on every run in order to capture the
case where an upstream submodule URL was changed. This doesn't happen often, but
it can be quite painful when it does, and fortunately 'git submodule sync' is a
local (and thus, cheap) command.